Remove Security and IAST statics in `AspNetCoreDiagnosticObserver` (#6564)
## Summary of changes

- Allow using a "local" instance of `Iast` in the
`AspNetCoreDiagnosticObserver` to reduce interactions across tests
- Remove some of the static accesses in these types

## Reason for change

We pass in an instance of the `Tracer` and `Security` into
`AspNetCoreDiagnosticObserver` to make it easier to test - accessing
`static` instances create cross-test interactions that are hard to debug
and hard to untangle.

However, _just_ moving the `Iast` to inject an instance isn't really
enough, because the non-static instances of `Security` and `Iast` access
global singleton values. That means there's easily a chance for
cross-talk between tests via these static instances. Ideally we should
refactor those to allow them to be optionally injected in the
constructor, and then to do that in the tests.

## Implementation details

- Optionally inject an instance of `Iast` into the diagnostic observer,
and use it preferentially
- Allow providing an `IDiscoveryService` to Iast, and use it if it's
provided
- Update tests to provide instances of `Security` and `Iast` to the
diagnostic observer
- Update benchmarks to use stub versions of the values to try to reduce
variation

## Test coverage

Coverage is the same