[ASM] Fix NullReferenceException on HttpTransport get_StatusCode (#6602)
## Summary of changes

Handle the errors when trying to access the `StatusCode` of the
`HttpResponse` from the Context.

## Reason for change

Errors in telemetry identified by these stacktraces:

```
System.NullReferenceException
   at Microsoft.AspNetCore.Http.DefaultHttpResponse.get_StatusCode()
   at Datadog.Trace.AppSec.Coordinator.SecurityCoordinator.RunWaf(Dictionary`2 args, Boolean lastWafCall, Boolean runWithEphemeral, Boolean isRasp)
```

A `ObjectDisposedException` and a `NullReferenceException` can happen in
concurrent access to the `HttpContext`.
This situation of race condition can happen while the request is
finished while async jobs still run in the background and the waf for
RASP is run.

The HttpContext can be disposed from different manner by the user. You
can find more information in this
[PR](https://github.com/DataDog/dd-trace-dotnet/pull/6529).

## Implementation details

- The getter for the `StatusCode` has been changed so it could return a
`int?` if an exception has been triggered.
- The method `ShouldAnalyzeSchema()` is already working with a null
`StatusCode`.