[Profiler] Implement outgoing HTTP requests profiling (#6280)
## Summary of changes
Add outgoing HTTP requests profiling to the .NET Continuous Profiler
- Compute the duration of DNS, socket, security, request/response phases
- sample based on duration and span ID

## Reason for change
Most HTTP request processing is done asynchronously. Due to the CPU/Wall
time profiling implementation, it is almost impossible to display them
in the timeline; leading to "empty" period of time. With the events
emitted by the BCL, it is possible to know which thread starts a request
and which one processes the response/error.

## Implementation details
Listen to HTTP related events from the BCL providers 

## Test coverage
- configuration
- enabled providers
- sample generation

## Other details
- fix remaining warnings

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->