Call `HttpRequest.BuildUrl` to avoid URL caching (#6361)
## Summary of changes

Manually compute the value of `HttpRequest.Url` for .NET Framework so
that that it won't be cached for other callers due to us.

## Reason for change

For .NET Framework we access `HttpRequest.Url` when adding data to
Spans.
When called this method will build the URL of the `HttpRequest` if we
are the first accessors and then the result will be cached in a private
field `_url`.

The issue here is that there seems to be a few edge cases where this
causes us to build the HttpRequest _before_ it is desired to be built.
 
## Implementation details

Adds a new feature flag: `BypassHttpRequestUrlCachingEnabled` to allow
bypassing of the current behavior where the URL would be cached.

Duck types the `HttpRequest` `BuildUrl` function to manually build the
URL to avoid caching it when enabled.

## Test coverage

Added some _unit tests_ for this that rely on reflection and attempted
to emulate previous behavior and the behavior we want to fix.

## Other details

I'm not confident that this a smart idea.

We will end up with an incorrect value here, but we did previously as
well (only in these rare cases) but we will hopefully get ourselves out
of the picture of affecting/altering application behavior.
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->

https://datadoghq.atlassian.net/browse/AIDM-500