Create an "installer" executable for use with the Datadog installer   (#6643)
## Summary of changes

Creates an "installer" executable, designed to be executed by the fleet
installer as part of configuring Windows SSI.

## Reason for change

We don't want to use the existing MSI, primarily because it requires
stopping and starting IIS when you update it, otherwise we risk causing
crashes.

The datadog-installer is responsible for copying the files added to the
Windows OCI image, and running the Datadog.FleetInstaller.exe executable
to configure the application. Similarly, it calls this exe when
uninstalling a tracer version or removing the product

## Implementation details

- Uses the fusion.dll API directly to manage the GAC entries, based on
the generated PInvoke APIs defined
[here](https://github.com/dotnet/pinvoke/tree/main/src/Fusion)
- Sets the envivornment variables by setting the `environmentVariables`
section in applicationHost.config. Uses [the
_Microsoft.Web.Administration_
NuGet](https://www.nuget.org/packages/Microsoft.Web.Administration) to
interact with the native API.

## Test coverage

Manually tested. Subsequent PRs in the stack will add smoke and
integration tests.

## Other details

Explicitly designed to run on Windows Server 2016+, so targets .NET FX
for simplicity. The _Microsoft.Web.Administration_ nuget is a .NET
Standard 1.x dll, which creates an annoying number of dlls in the
output, but it's just ugly, it works fine.

There's a fair amount of duplication in the Fusion
API/Microsoft.Web.Administration code with what the
`dd-dotnet`/`dd-trace` tool currently does. We could certainly look at
consolidating that down the line, and potentially rewriting the fleet
installer tool to be a NativeAOT executable, but it's probably not worth
the effort at this point.

The explicit `<RuntimeIdentifier>win-x64</RuntimeIdentifier>` in the
_.csproj_ is an odd one. Without it, for some reason, the gitlab build
restores for x86 but tries to build for x64 which is... weird. Also,
can't repro the issue locally. _Really_ don't understand what's
happening there, but this works so... 🙈

Part of a stack of PRs:
 
- https://github.com/DataDog/dd-trace-dotnet/pull/6643 👈
- https://github.com/DataDog/dd-trace-dotnet/pull/6644
- https://github.com/DataDog/dd-trace-dotnet/pull/6645