Build the FleetInstaller executable and include in Windows OCI image (#6644)
## Summary of changes

- Build the fleet installer exe in GitLab
- Include the fleet installer exe in Windows OCI
- Produce Windows OCI images in Staging

## Reason for change

We need to build the fleet installer executable, and include it in a
Windows OI image

## Implementation details

- Add a stage `BuildFleetInstaller` to explicitly publish the fleet
installer
- Update the `download-single-step-artifacts` to grab the
`windows-tracer-home.zip` and fleet installer artifact from the `Build`
stage
- Update the OCI image

For tagged releases this flow is a bit more convoluted. We need to grab
the fleet-instller.exe artifact from the s3 upload that's done in the
`publish` step, because we don't store it on the GitHub release. It's a
bit ugly, but it works.

## Test coverage

This is the test - if it builds, and the OCI image contains the expected
files, we're good

## Other details

Note that _currently_, the final OCI images will _not_ contain Windows
layers, until we unblock that both in our gitlab YAML, and in the
libdatadog-build one pipeline.


Part of a stack of PRs:
 
- https://github.com/DataDog/dd-trace-dotnet/pull/6643
- https://github.com/DataDog/dd-trace-dotnet/pull/6644 👈
- https://github.com/DataDog/dd-trace-dotnet/pull/6645