Add smoke tests for fleet installer (#6645)
## Summary of changes

Adds a new smoke test stage, for running inside Windows using the fleet
installer

## Reason for change

Emulates an end-to-end smoke test for the fleet installer executable. 

## Implementation details

- Build the fleet installer in the AzDo pipeline, as part of the
`package-windows` stage
- Adds a new `fleet_installer_smoke_tests` stage for running the smoke
tests
- Fix issues in ASP.NET Core test app, so that it works when hosted in
IIS
- Add a dockerfile for running the ASP.NET Core app with the fleet
installer, hosted in IIS

That latter point proved to be somewhat of a nightmare. The resulting
dockerfile feels kind of horrible, but I couldn't find another way to
have:

- Delay the start of the app pool until the container image is running
(i.e. not during build time)
- Start the worker process when the app pool starts, _without_ an
incoming request
- Shut down the app pool when the worker process exits (which is the way
the aspnetcore app works)
- Exit the container

The workaround, using the entrypoint script, makes a 404 request to the
app to spin up the worker process, manually shuts down the app pool (to
make sure the app definitely shuts down and flushes), and then exits.
This causes an additional span in the traces, so needed to create new
snapshots.

## Test coverage

This gives basic snapshot testing on 2022 which is the only version I
could get to work given the hosted images on AzDo and the available
python base images for the test agent. I think that is sufficient - we
are going to have additional end-to-end testing of Windows images in
system tests anyway.

## Other details

Part of a stack of PRs:

 
- https://github.com/DataDog/dd-trace-dotnet/pull/6643
- https://github.com/DataDog/dd-trace-dotnet/pull/6644
- https://github.com/DataDog/dd-trace-dotnet/pull/6645 👈