Fix `DisabledIntegrationNames` not correctly disabling the integration (#6664)
## Summary of changes

Fixes the case where `DisabledIntegrationNames` is modified in code but
ignored.

## Reason for change

This code:

```csharp
var settings = TracerSettings.FromDefaultSources();
settings.DisabledIntegrationNames.Add("Kafka");
Tracer.Configure(settings);
```

does _not_ cause the Kafka integration to be disabled as it should be.

## Implementation details

We were missing a remapping. The `IConfigurationSource` expects to read
objects as strings, so when we transfer from manual <> automatic we have
to convert the provided `HashSet<string>` to a string (to then later be
converted _back_ to a `HashSet<>`&hellip;I know, I know). This
conversion was missing, the calling code was expecting a `string`, so it
failed.

## Test coverage

Updated the unit tests to demonstrates the breaking behaviour (this was
a missing part of the tests, because the mapping with
`DisabledIntegrationNames` and `Integrations` is a bit of a confusing
mess unfortunately). After the fix, the tests pass.

## Other details

Fixes #6657

There are multiple workarounds to this behaviour currently. You can set
env vars like this;

- `DD_TRACE_Kafka_Enabled="false"`
- `DD_DISABLED_INTEGRATIONS="Kafka"`

Or you can use the `Integrations` property:

```csharp
var settings = TracerSettings.FromDefaultSources();
settings.Integrations["Kafka"].Enabled = false;
Tracer.Configure(settings);
```