Tack on apm-idm-dotnet reviewers to auto-bump PRs (#6665)
## Summary of changes

Adds `apm-idm-dotnet` as a reviewer to auto PRs. 

## Reason for change

I get a summary of open PRs, but these are missing

## Implementation details

Added the teams
Updated to `team-reviewers` as that seems to be more correct according
to documentation

## Test coverage

## Other details
<!-- Fixes #{issue} -->

Note: this shouldn't make `apm-idm-dotnet` a _required_ reviewer. If it
does then revert this change.

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->