Update `CONTRIBUTING` with external PR policies (#6650)
## Summary of changes

Updates `CONTRIBUTING` to be more transparent to contributors about how
external PRs need to be made and that they'll need to be re-created by a
core contributor.

Also expanded on new library and updated version to point to the
helpdesk as that is the quickest route usually.

## Reason for change

These policies / limitations have been in place for a long time, but not
well communicated.

Attempting to be more transparent

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->