Try to catch memory dump for instrumentation telemetry tests (#6675)
## Summary of changes

Try to grab a memory dump if `dotnet build` fails as part of the
instrumentation tests

## Reason for change

Since updating to 9.0.102, we've seen cases of the `dotnet build`
hanging on arm64. In this test we explicitly build a "forwarder" app,
and that has a hung several times in CI now.

## Implementation details

Use the memory dump helper to try to grab a dump. We've only seen this
happen on linux arm64 so far, so haven't bothered to hook up the Windows
monitoring or anything like that

## Test coverage

If the tests pass, that's good enough for me until we see this flake
manifest again.