[Fleet Installer] Don't block uninstall if IIS isn't available (#6676)
## Summary of changes

If IIS is uninstalled _before_ you run the `uninstall-product` command,
don't fail the uninstall.

## Reason for change

It's a niche issue. We block install if IIS isn't available or is <10.0,
but we currently also block the _uninstall_. This PR makes it so that if
the customer uninstalls IIS after installing the product, we don't block
the uninstall

## Implementation details

If they uninstalled IIS, there's not really anything _to_ uninstall, as
this stage just removes the env vars from the apps (which isn't possible
once IIS is uninstalled). So basically just move the check from being a
global check to an install-only check.

## Test coverage

:shhh: