Pack trimming file in Datadog.Trace NuGet, and add trimming smoke tests for the NuGet packages (#6678)
## Summary of changes

- Pack the trimming file into the _Datadog.Trace_ NuGet package
- Add smoke tests for the _Datadog.Trace_ and _Datadog.Trace.Trimming_
NuGet packages
- Remove the unnecessary sdk.props file from _Datadog.Trace.Trimming_

## Reason for change

When users are building trimmed (non-aot) apps, they need to reference
our trimming file, which is currently exposed via the
_Datadog.Trace.Trimming_ NuGet package. We created this as a separate
package, because we didn't want people to have to reference the
_Datadog.Trace_ NuGet if they didn't need it.

However, there's not really a good reason to force people to _have_ to
reference another NuGet package if they are _already_ referencing
_Datadog.Trace_. And there's no harm in including the trimming file for
non-trimmed apps, it has no impact.

## Implementation details

- Pack the same trimming.XML file into the _Datadog.Trace_ NuGet package
- Include a _Datadog.Trace.props_ file to automatically reference the
trimmer file when the NuGet is referenced
- Added smoke tests that confirm the _NuGet_ packaging is correct for
both the new package, and the existing package
- We already had "integration" tests that confirm the trimmer file
works, but we didn't have any tests that were using the NuGet package,
to confirm the `props` are correctly added etc.
- Removed the `sdk/sdk.props` file from _Datadog.Trace.Trimming_ as it's
unused. This file is only useful if you're creating a "[custom .NET
SDK](https://learn.microsoft.com/en-us/visualstudio/msbuild/how-to-use-project-sdk?view=vs-2022#use-the-sdk-attribute-on-the-project-element)"
to reference in the `<Project SDK="Some.SDK">` element (which we're
not).

## Test coverage

Added smoke tests to confirm both the _Datadog.Trace_ and
_Datadog.Trace.Trimming_ packages work as expected. For now, these are:

- `linux-x64`/`linux-musl-x64` only
- `net8.0`/`net9.0` only

We could expand that further if we want, but I'm not massively inclined
to do so - we primarily want to ensure the _packages_ are valid, and we
don't really need to run that across all the different platforms IMO.
Plus it will just increase the chance of flake.

## Other details

As part of this, noticed that crash tracking didn't work. 
- Disabled crash tracking in these tests for now
- https://github.com/DataDog/dd-trace-dotnet/pull/6677 was raised in
response
- Ideally we should merge this, and then re-enable the crash tracking
tests in that PR

I've kept these in the "standard, every PR" set of smoke tests for now,
but maybe we should move them to the master-only set? 🤷‍♂️

---------

Co-authored-by: Kevin Gosse <kevin.gosse@datadoghq.com>