speed up external CI artifact fetching (#6674)
## Summary of changes
Making an artifact out of Datadog.Trace.Manual.dll
## Reason for change
As part of migrating the throughput benchmarks to Gitlab, we need
"Datadog.Trace.Manual.dll" and using the AzDo API we are only able to
fetch whole artifacts not files within it. The
build-windows-working-directory takes a considerable amount of time to
download.
## Implementation details
Publishing the artifact
## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->