[ASM][ATO] Collect session id at all times (#6623)
## Summary of changes

Collects session id in any case:
- .net core: on endpoint matched
- net framework: try access it through httpcontext
User doesn't need to be authenticated anymore for it to collect the
session id.

## Reason for change

Session id should always be collected, not only when user is
authenticated.

## Implementation details

there are guards for netcore as accessing Session when it's not been
setup just throws. So this way we make sure it's setup for the webapp.

## Test coverage

Change all snapshots.
Scrubbing pitfalls: 
- scrub cookies and session id for aspnet mvc because if there's a
session id it will always change the cookies
- scrub cookies and session id for aspnet core when user is
authenticated / logged in (as it will keep changing)
- other cases in aspnet core: just scrub session fingerprint
## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->