Add workflows for scheduled code freeze and AAS deployment (#6681)
## Summary of changes

This adds three new workflows to support a scheduled code freeze and
test app deployment every _even_ week number on Friday afternoon EST
time.

## Reason for change

Think it'd be better to more toward a more proactive approach for code
freezing and releasing.

## Implementation details

- `scheduled_code_freeze.yml`
  - Runs every Friday at 4PM EST
  - Odd week -> skipped freeze
  - `Skip Scheduled Code Freeze` is `open` -> skipped freeze
  - Even week -> freeze
    - Update PR status
- `scheduled_aas_deploy.yml`
  - Runs every Friday at 4:30PM EST
- Unsure how to trigger it to be ready when the AzDo stage has built the
necessary artifacts, so this is hacky
    - Odd week -> skipped deploy
    - `Skip Scheduled Code Freeze` is `open` -> skipped deploy
      - close the `Skip Scheduled Code Freeze`
    - Even week -> deploy
- `create_skip_code_freeze.yml`
  - Opens `Skip Scheduled Code Freeze`
- `pr-status-updater/action.yml`
- Copy / modification of `code-freeze/action.yml` to be configurable,
but haven't tested this with large number of PRs
- Skip Scheduled Code Freeze Milestone:
https://github.com/DataDog/dd-trace-dotnet/milestone/183

## Test coverage

I testing this out in a different repository as that was easier / safer.

Let me know if you want access to that :)

The `pr-status-updater/action.yml` is a version of
`code-freeze/action.yml` but more configurable, only thing is that I
needed to use `curl` instead of the GitHub action and I didn't test on
the same number of PRs. **Let me know if it makes more sense to swap to
`code-freeze/action.yml`**

I left in the workflow dispatch stuff as it is easier to just leave them
in for testing purposes

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>