Add NuGet API verification check (#6690)
## Summary of changes

Adds an early stage in the create_draft_release workflow to check that
the Nuget API key is valid

## Reason for change

We recently did a release, and the "upload NuGet packages step failed".
Unfortunately, this happens _after_ the draft release is created which
is essentially a one-way trip (it triggers downstream things).

## Implementation details

Unfortunately, due to 
- https://github.com/NuGet/Home/issues/13475
- https://github.com/NuGet/NuGetGallery/issues/8283

There's no good way to test the key _Without_ pushing a package.

So this takes a hacky approach:
- Download one of our packages
- Try to push the package (and set `--skip-duplicate` so that it becomes
a no-op)
- If it succeeds, it's a no-op, if the key is invalid, it fails, and we
bail

## Test coverage

Tested the steps locally, and they work so 🤞