Add ability to disable `ActivitySource`s from being listened to (#5795)
## Summary of changes

This adds the ability for `ActivitySource`s to be disabled so that they
won't be listened to by the .NET Tracer.

## Reason for change

Currently there is no way to disable an `ActivitySource` if
OpenTelemetry is enabled in the .NET Tracer without entirely disabling
OpenTelemetry support. The `IgnoreActivityHandler` is insufficient as it
will still _listen_ to the `ActivitySource` so they will create
`Activity` objects.

## Implementation details

- Adds `DD_TRACE_DISABLED_ACTIVITY_SOURCES` that accepts a
comma-separated list of `ActivitySource` names that supports the
glob-pattern.
- example: `"DD_TRACE_DISABLED_ACTIVITY_SOURCES":
"SomeSourceName,*SomeGlobPattern*"`
- Adds `DisableActivityHandler` that will determine based on the above
environment variable whether or not a given `ActivitySource` should be
listened to.

## Test coverage

- Updated the `NetActivitySdk` sample project with an ignored
`ActivitySource` and some that are disabled.

## Other details
<!-- Fixes #{issue} -->

When merged, get this merged:
https://github.com/DataDog/dd-go/pull/171612

Fixes #5740 

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->