Add fixes for release process (#6704)
## Summary of changes

- Increase HttpClient download timeout for Azure Devops release
artifacts
- Make sure destination temp directory has been created
- Fix issue in GitLab release pipeline for Windows

## Reason for change

Saw failures due to all of these issues in latest release.

## Implementation details

Increase the timeout to download the release artifacts from "ridiculous"
(100s) to "ludicrous" (6000s), in the hope that we don't hit it again.
It's safe to just retry in the UI here anyway, so this will hopefully be
sufficient.

Additionally, a recent additional check (the fleet installer) was
failing because the target directory for the download hadn't been
created, so fixed that.

Finally, the `3.11.0` release is broken because of a bug in the script,
which only manifests when running for a tag

## Test coverage

Ran the Build stage locally - they both failed previously and now both
pass.

We can't easily test the gitlab change unfortunately