Add `DD_INSTRUMENTATION_INSTALL_TYPE` variables for known install methods (#6695)
## Summary of changes

Add the `DD_INSTRUMENTATION_INSTALL_TYPE` for install types for which we
can set it

## Reason for change

The `DD_INSTRUMENTATION_INSTALL_TYPE` is used to record when ssi
instrumentation is used (for example). I think it makes sense to set
this value for some of our installation methods (the ones that set
variables automatically for users)

## Implementation details

Add `DD_INSTRUMENTATION_INSTALL_TYPE` for:
- The windows fleet installer (`windows_fleet_installer`)
- The MSI (`dotnet_msi`)
- `dd-trace` tool (`dd_trace_tool`)
- `dd-dotnet` launcher (`dd_dotnet_launcher`)

## Test coverage

Added a small test to the `dd-dotnet` test (because it was easy, and we
were already testing this sort of thing). I don't think it's worth the
effort to automatically test the rest personally, but if _someone_ wants
to ruin my day they have a great excuse.

## Other details

Note that `dd-dotnet` _doesn't_ use a get env var wrapper, but that's
by-design as it doesn't run in partial trust, and there are bigger
problems if that fails!