[Fleet installer] Add additional pre-verification check for IIS variables (#6697)
## Summary of changes

- Add additional pre-install check for whether we can read IIS config
before doing anything irreversible
- If the variables have not been set, then if changing the variables
fails, try to revert the changes

## Reason for change

Where possible, we want to revert actions safely. On the first install,
if modifying IIS fails, we could theoretically end up with a strange
state in their IIS config. This attempts to revert the changes if we
think it's safe to do so.

## Implementation details

Tries to look for the presence of the `DD_INSTRUMENTATION_INSTALL_TYPE`
variable with value `windows_fleet_installer` - if we find it, then we
consider this to be an "update" rather than an "initial install". We
_don't_ want to rollback the IIS changes in case of an update, so we
leave the variables.

## Test coverage

Yeah... I still need to add integration tests

## Other details

With this design, there's still a risk that we remove variables when we
_shouldn't_. e.g. if the customer has _manually_ added instrumentation
to the app pools using the same method, we would delete these variables.

Similarly, if they have explicitly opted out of instrumentation, by
setting `COR_ENABLE_PROFILING=0` for example, then this would be
removed. But currently we _already_ stomp on these (which is something
else for us to think about behaviour-wise)

Requires
- https://github.com/DataDog/dd-trace-dotnet/pull/6695