Fix: Ensure baggage properties are not interpreted as baggage values (#6693)
Fixes an issue where baggage properties were being interpreted as part
of a baggage value (each baggage item consists of a key, a value, and
properties). The OpenTelemetry Baggage API does not include support for
properties, so the decision was made to exclude the propagation of
properties, so this change discards them when extracting them from the
header.