Fix "auto bump test package versions" workflow  (#6713)
## Summary of changes

- Fixes the broken workflow
- Updates the version of `peter-evans/create-pull-request` (as a safe
test of the newer version)

## Reason for change

https://github.com/DataDog/dd-trace-dotnet/pull/6665 tried to
automatically add reviewers to the auto-created PR, but tl;dr; it didn't
work, and was causing the action to fail. We could look at using a PAT
to work around the issue, but IMO it's not worth it, so reverted.

As I was testing it anyway, also bumped the version of the action - we
can look at bumping it everywhere if this doesn't run into issues.

## Implementation details

Remove the `team-reviewers` option and manually update the package
version

## Test coverage

Did a test run
[here](https://github.com/DataDog/dd-trace-dotnet/actions/runs/13519784696/job/37776296046)
and it worked

## Other details

Partially reverts https://github.com/DataDog/dd-trace-dotnet/pull/6665