Perf optimization for protobuf instrumentation (#6694)
## Summary of changes

Stop relying on the proto source file to know if we're dealing with an
internal protobuf message.
Instead, we check the type at the very beginning, and disable the
instrumentation for that type if we detect it's a google type.
This means we can also get rid of all the dark magic we had around
making sure we can access the descriptor, because we now know we'll only
do it for regular messages.

## Reason for change

it's better perf

## Implementation details

I use an other `OnMethodXxx` method to be able to access the raw type, I
explained it in the comments.
See [this
thread](https://dd.slack.com/archives/CC7Q58YRG/p1740062564566959) for
more.

## Test coverage

added a path in the integration tests that checks that

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->

---------

Co-authored-by: Lucas Pimentel <lucas.pimentel@datadoghq.com>