don't propagate baggage when http client integration is disabled (#6692)
## Summary of changes

When implementing baggage in #6157., we changed the logic in
`HttpMessageHandlerCommon` so that it would propagate baggage even if
the http client integration was disabled. This changes this behavior so
that headers are only injected if the integration is enabled.

Bonus change: cleaned up the integration tests that were modified in
this PR
- added `StringUtil.GetAllHeaders()` so we only run the `Regex` once on
the stdout instead of a separate regex for each header
- use `FluentAssertions`
- use modern C# constructs like `var` and `using` statements (I
recommend ignoring whitespace when viewing the diff)

## Reason for change

This behavior is more correct. Disabling the integration should mean the
integration doesn't change any app behavior at all, including injecting
http headers. There are separate settings to disable trace and baggage
propagators, if needed.

## Implementation details

Move the `Inject()` call into the `if(enabled)` block, where it used to
be before #6157.

## Test coverage

- Added assertions to
`HttpMessageHandlerTests.TracingDisabled_DoesNotSubmitsTraces()` to
verify that _none_ of the known propagation headers are injected when
integration is disabled (not just `baggage`).

<!--
## Other details
-->

<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->