[Test Optimization] Add missing impacted test code in runner (#6722)
## Summary of changes

This PR adds some missing code in the dd-trace runner related to
impacted test feature.

## Reason for change

We can avoid making some requests to the backend if we have all the
information we need (calculated in the runner). Impacted tests
configuration was missing in the runner, so we had to go to the settings
api in every testhost process. This PR fixes that.

## Implementation details

Just add the missing code like the other features.

## Test coverage

This is tested in the `test-environment` repo.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->