[Test Optimization] Add known tests feature (#6725)
## Summary of changes

This PR adds support for Known Tests feature as described in the
following RFC:
https://docs.google.com/document/d/1mPMM7Uye2jHUMx5ADTMSRt_Rff98klRLj5OpTruszzQ/edit?pli=1&tab=t.0

https://datadoghq.atlassian.net/browse/SDTEST-1277

## Reason for change

Is a new feature of the test optimization library.

## Test coverage

We updated the EFD tests to include the retry reason in the snapshots

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->