[Profiler] Properly initialize appDomainId (#6630)
## Summary of changes

The contention profiler and the allocation profiler don't set the
appDomainId if the managed thread can't be found (it can happen, for
instance, if it has exited). Because the `RawSample` isn't zero'd at
construction, this causes the field to contain garbage, which can cause
crashes down the line.

Also, some error paths set the appDomainId to -1. I changed it to 0
because it's the value expected by the code that uses it and by the CLR.

## Reason for change

This was found thanks to a crash dump in the CI, but there has been
crash reports that we failed to understand until now.