Fix HttpListener.GetContext leak in tests (#6750)
## Summary of changes

Replace calls to `HttpListener.GetContext` with
`HttpListener.GetContextAsync`

## Reason for change

On .NET Core 3.1, the tests were leaking threads because of an issue in
`HttpListener.GetContext`, where the call to `GetContext` doesn't end
even after disposing the listener:
https://github.com/dotnet/runtime/issues/25497

Apparently, the async APIs is not affected by this bug.

## Test coverage

Tested on my machine. I observed the thread increase and the OOM before
the change. Everything seems fine after the change.

## Other details

Also fixed a missing dispose for one `MockTracerAgent`.