[fleet installer] Add some integration tests (#6714)
## Summary of changes

Adds some "integration" tests for the fleet installer component

## Reason for change

We don't have a lot of in-repo tests for the installer currently.
Creating sub-cutaneous and integration tests allow more fine-grained
testing, particularly of edge cases.

## Implementation details

The fleet installer messes with fundamental components like the GAC so
running these tests in a specific separate job (in the MSI installer
tests stage, seemed like the right place for it)

Currently these tests are focusing on the lower-level components like
the `GacHelper` and `RegistryHelper` etc. I want to add similar levels
of testing for the IIS integration (`AppHostHelper`) as well as some
full/subcutaneous tests for the install/uninstall commands, I just ran
out of time.

## Test coverage

More coverage now, but we could still do more. The GAC tests which don't
yet pass are skipped for now.

## Other details

One aspect that this flagged, is that currently the GAC uninstall does
_not_ behave as we originally believed (based on initial testing and
`gacutil`'s behaviour). Currently we're failing to remove entries when
there's already an entry with the same name (i.e. everytime we do an
upgrade...)

We're currently treating the error
`IASSEMBLYCACHE_UNINSTALL_DISPOSITION_REFERENCE_NOT_FOUND` as a
"success" case, as it looks like the file has already been removed. But
in reality, we're actually not finding the assembly to remove it when we
should.

When there's a _single_ assembly in the GAC, there's no issue, and we
can use the "simple" name, e.g. `Datadog.Trace` to uninstall the
library. But AFAICT when there's multiple, we need the "full" name, e.g.
`Datadog.Trace, Version=3.12.0.0, Culture=neutral,
PublicKeyToken=def86d061d0d2eeb, processorArchitecture=MSIL`

(None of this is documented, just what I'm observing in my tests)
(Also note that `gacutil.exe` behaves differently - it will happily use
just a `Datadog.Trace` and (AFAICT) enumerates the assemblies and tries
to remove each of them)